daemon: Fix unit test panic in daemon init failure
The below panic was seen during TearDownTest() when the daemon init
function fails. We can avoid the panic and fail out gracefully by
checking that the daemon was instantiated before dereferencing it.

----------------------------------------------------------------------
PANIC: labels_test.go:89: DaemonSuite.TearDownTest

... Panic: runtime error: invalid memory address or nil pointer dereference (PC=0x459388)

/usr/local/go/src/runtime/panic.go:489
  in gopanic
/usr/local/go/src/runtime/panic.go:63
  in panicmem
/usr/local/go/src/runtime/signal_unix.go:290
  in sigpanic
labels_test.go:90
  in DaemonSuite.TearDownTest
/usr/local/go/src/reflect/value.go:302
  in Value.Call
/usr/local/go/src/runtime/asm_amd64.s:2197
  in goexit

----------------------------------------------------------------------
PANIC: state_test.go:175: DaemonSuite.TestCleanUpDockerDangling

... Panic: Fixture has panicked (see related PANIC)

----------------------------------------------------------------------

+++ daemon/labels_test.go
@@ -87,7 +87,9 @@ func (ds *DaemonSuite) SetUpTest(c *C) {
 }
 
 func (ds *DaemonSuite) TearDownTest(c *C) {
-	os.RemoveAll(ds.d.conf.RunDir)
+	if ds.d != nil {
+		os.RemoveAll(ds.d.conf.RunDir)
+	}
 }
 
 func (ds *DaemonSuite) TestLabels(c *C) {
